5.28

+++ BackEnd/Co-Design-Connect/src/main/java/org/example/codesignconnect/model/File.java
@@ -12,8 +12,8 @@ import java.sql.Timestamp;
 public class File {
     private Integer id;
     private Integer projectId;
-    private Byte projectStatus;
-    private Byte type;
+    private Short projectStatus;
+    private Short type;
     private Integer iterationId;
     private String name;
     private String source;
